feat(unionlabs): move types in tendermint/types to their own files
also changes CanonicalVote's field to use the correct types
