+++ lib/unionlabs/src/tendermint/types/canonical_block_header.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+use crate::ethereum::H256;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct CanonicalPartSetHeader {
+    pub total: u32,
+    pub hash: H256,
+}
