+++ lib/unionlabs/src/tendermint/types/canonical_block_id.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+use crate::{ethereum::H256, tendermint::types::canonical_block_header::CanonicalPartSetHeader};
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct CanonicalBlockId {
+    pub hash: H256,
+    pub part_set_header: CanonicalPartSetHeader,
+}
