+++ lib/unionlabs/src/tendermint/types/signed_msg_type.rs
@@ -1,0 +1,11 @@
+use crate::macros::wrapper_enum;
+
+wrapper_enum! {
+    #[proto(protos::tendermint::types::SignedMsgType)]
+    pub enum SignedMsgType {
+        Unknown = 0,
+        Prevote = 1,
+        Precommit = 2,
+        Proposal = 32,
+    }
+}
